refactor(cli): retain help projection boundary
